+++ package.json
@@ -49,6 +49,7 @@
     "@nestjs/platform-express": "^11",
     "@nestjs/swagger": "^11.1.1",
     "@nestjs/terminus": "^11",
+    "@nestjs/throttler": "^6.5.0",
     "@types/json-merge-patch": "^1.0.0",
     "@types/jsonschema": "^1.1.1",
     "@user-office-software/duo-logger": "^2.1.1",
