+++ src/app.module.ts
@@ -44,6 +44,8 @@ import { HistoryModule } from "./history/history.module";
 import { MaskSensitiveDataInterceptorModule } from "./common/interceptors/mask-sensitive-data.interceptor";
 import { RuntimeConfigModule } from "./config/runtime-config/runtime-config.module";
 import { MetadataKeysModule } from "./metadata-keys/metadatakeys.module";
+import { OidcClientModule } from "./common/openid-client/openid-client.module";
+import { ThrottlerModule } from "@nestjs/throttler";
 
 @Module({
   imports: [
@@ -53,6 +55,7 @@ import { MetadataKeysModule } from "./metadata-keys/metadatakeys.module";
       cache: true,
     }),
     AuthModule,
+    OidcClientModule,
     CaslModule,
     AttachmentsModule,
     CommonModule,
@@ -164,6 +167,15 @@ import { MetadataKeysModule } from "./metadata-keys/metadatakeys.module";
       MaskSensitiveDataInterceptorModule,
       (env: NodeJS.ProcessEnv) => env.MASK_PERSONAL_INFO === "yes",
     ),
+    ThrottlerModule.forRoot({
+      throttlers: [
+        {
+          name: "login",
+          ttl: 1000,
+          limit: 1,
+        },
+      ],
+    }),
   ],
   controllers: [],
   providers: [
