+++ src/auth/auth.module.ts
@@ -8,22 +8,20 @@ import { JwtModule } from "@nestjs/jwt";
 import { JwtStrategy } from "./strategies/jwt.strategy";
 import { LdapStrategy } from "./strategies/ldap.strategy";
 import { ConfigService } from "@nestjs/config";
-import { UsersService } from "src/users/users.service";
 import { OidcConfig } from "src/config/configuration";
 import { OidcStrategy } from "./strategies/oidc.strategy";
-import { accessGroupServiceFactory } from "./providers/access-group-provider/access-group-service-factory";
-import { AccessGroupService } from "./providers/access-group-provider/access-group.service";
+import { accessGroupServiceFactory } from "./access-group-provider/access-group-service-factory";
 import { CaslModule } from "src/casl/casl.module";
 import { SessionMiddleware } from "./middlewares/session.middleware";
-import { OidcClientService } from "./providers/openid-cilent.service";
+import { OidcClientService } from "../common/openid-client/openid-cilent.service";
+import { OidcAuthService } from "src/common/openid-client/openid-auth.service";
 
 const OidcStrategyFactory = {
   provide: "OidcStrategy",
   useFactory: async (
     oidcClientService: OidcClientService,
+    oidcAuthService: OidcAuthService,
     configService: ConfigService,
-    userService: UsersService,
-    accessGroupService: AccessGroupService,
   ) => {
     if (!configService.get<OidcConfig>("oidc")?.issuer) {
       return null;
@@ -31,15 +29,10 @@ const OidcStrategyFactory = {
 
     const client = await oidcClientService.getClient();
 
-    const strategy = new OidcStrategy(
-      client,
-      configService,
-      userService,
-      accessGroupService,
-    );
+    const strategy = new OidcStrategy(client, configService, oidcAuthService);
     return strategy;
   },
-  inject: [OidcClientService, ConfigService, UsersService, AccessGroupService],
+  inject: [OidcClientService, OidcAuthService, ConfigService],
 };
 
 @Module({
@@ -60,7 +53,6 @@ const OidcStrategyFactory = {
     UsersModule,
   ],
   providers: [
-    OidcClientService,
     AuthService,
     JwtStrategy,
     LdapStrategy,
