+++ src/auth/strategies/ldap.strategy.ts
@@ -7,7 +7,7 @@ import { CreateUserDto } from "src/users/dto/create-user.dto";
 import { CreateUserIdentityDto } from "src/users/dto/create-user-identity.dto";
 import { FilterQuery } from "mongoose";
 import { User, UserDocument } from "src/users/schemas/user.schema";
-import { AccessGroupService } from "../providers/access-group-provider/access-group.service";
+import { AccessGroupService } from "../access-group-provider/access-group.service";
 import { UserPayload } from "../interfaces/userPayload.interface";
 import { Profile } from "passport";
 import { UserProfile } from "src/users/schemas/user-profile.schema";
@@ -16,7 +16,7 @@ import { LdapConfig } from "src/config/configuration";
 @Injectable()
 export class LdapStrategy extends PassportStrategy(Strategy, "ldap") {
   constructor(
-    private configService: ConfigService,
+    configService: ConfigService,
     private usersService: UsersService,
     private accessGroupService: AccessGroupService,
   ) {
